Update logging, scheduling config

+++ event_processor/entrypoint.sh
@@ -1,8 +1,15 @@
-rm -f twistd.pid
-scrapyd & 
-rm -rf eggs
-rm -rf logs
-rm -rf dbs
-chmod +x deploy.sh
-sleep 5
-./node_modules/nodemon/bin/nodemon.js -L --exec ./deploy.sh
+if [ "$RUN_SCHEDULER" = "1" ]
+then
+    rm -f twistd.pid
+    scrapyd & 
+    rm -rf eggs
+    rm -rf logs
+    rm -rf dbs
+    chmod +x deploy.sh
+    # -u is passed to python to return output unbuffered, so there is no delay when viewing docker logs
+    python -u create_schedules.py
+    sleep 5
+    scrapyd-deploy
+else
+    python -u runner.py
+fi
